fix: ajustes simples
- Renomeado nome de arquivo principal de Azure BS
- Ajuste para usar context externo  Azure BS
- Removendo número mágicos
- RedisDB ajuste no status ao gravar dado
- Add pasta de exemplos
